third_party/WebKit/LayoutTests/fast/events/onunload-window-property-expected.txt

CONSOLE ERROR: line 13: Blocked alert('unload') during unload.
You should have seen an unload alert appear.
